+++ internal/securefile/platform_other.go
@@ -1,0 +1,11 @@
+//go:build !windows
+
+package securefile
+
+var validatePlatformFilePermissions = func(string) error {
+	return nil
+}
+
+var validatePlatformParentDirectoryPermissions = func(string) error {
+	return nil
+}